requirements.txt
requests
google-cloud-bigquery
python-dotenv
pytest
python2
pandas
google-cloud-logging